Removed NoSuchAlgorithmException and changed the String role to Enum inside CreateUserRequest
