+++ src/main/java/com/project/trackfit/core/ApplicationUserController.java
@@ -8,7 +8,6 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.security.NoSuchAlgorithmException;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.UUID;
@@ -22,19 +21,18 @@ import static org.springframework.http.HttpStatus.CREATED;
 @RequestMapping("/api/v1/auth/register")
 public class ApplicationUserController {
 
-    private final ApplicationUserService applicationUserService;
+    private final IApplicationUserService IApplicationUserService;
 
     @PostMapping
-    public ResponseEntity<APICustomResponse> createUser(@Valid @RequestBody CreateUserRequest createUserRequest)
-            throws NoSuchAlgorithmException {
-        UUID userId = applicationUserService.createUser(createUserRequest);
+    public ResponseEntity<APICustomResponse> createUser(@Valid @RequestBody CreateUserRequest createUserRequest) {
+        UUID userId = IApplicationUserService.createUser(createUserRequest);
         Map<String, UUID> data = new HashMap<>();
         data.put("User_Id", userId);
         return new ResponseEntity(
                 APICustomResponse.builder()
                         .timeStamp(now())
                         .data(data)
-                        .message("Personal Trainer have been Created Successfully")
+                        .message("Application user has been created successfully")
                         .status(CREATED)
                         .statusCode(CREATED.value())
                         .build(),
