+++ src/main/java/com/project/trackfit/core/IApplicationUserService.java
@@ -1,0 +1,8 @@
+package com.project.trackfit.core;
+
+import java.security.NoSuchAlgorithmException;
+import java.util.UUID;
+
+public interface IApplicationUserService {
+    UUID createUser(CreateUserRequest createUserRequest);
+}
